PyAI_onLINE/EJEMPLOS/filtros_webcam/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
altair==5.5.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bce-python-sdk==0.9.56
blinker==1.9.0
branca==0.8.2
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comtypes==1.4.14
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==6.0
dlib @ file:///C:/Users/torre/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
easyocr==1.7.2
exceptiongroup==1.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.20.1
filterpy==1.4.5
Flask==3.1.2
flatbuffers==25.12.19
folium==0.16.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.4.2
ninja==1.13.0
nodejs-wheel-binaries==24.12.0
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pefile==2024.8.26
pillow==10.4.0
plotly==5.13.0
polars==1.36.1
polars-runtime-32==1.36.1
prettytable==3.17.0
propcache==0.4.1
protobuf==3.19.6
psutil==7.1.3
py-cpuinfo==9.0.0
pyairtable==3.3.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.5
pypdfium2==5.2.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pyserial==3.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytubefix==10.3.6
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.1
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
srt==3.5.3
streamlit==1.33.0
streamlit_folium==0.20.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tifffile==2025.5.10
tk==0.1.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
ultralytics==8.3.243
ultralytics-thop==2.0.18
urllib3==2.6.2
vosk==0.3.45
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xyzservices==2025.11.0
yarl==1.22.0
zope.interface==8.1.1
